==> california-housing/notebooks/1-eda-with-pyspark.ipynb <==
# Exploratory Data Analysis
In this notebook, we conduct some basic EDA on the 'California Housing' data set using PySpark. We are using a single machine to test out PySpark functionality for now.

### Configure Spark

# Import SparkSession
from pyspark.sql import SparkSession

# Build the SparkSession
spark = SparkSession.builder \
   .master("local") \
   .appName("Linear Regression Model") \
   .config("spark.executor.memory", "1gb") \
   .getOrCreate()
   
sc = spark.sparkContext

### Load and Examine Data
> run `bash download_data.sh` in the california-housing directory to retrieve the data from the web

We begin by interacting with the data as an RDD (resilient distributed dataset).

# load in the data
rdd = sc.textFile('../data/CaliforniaHousing/cal_housing.data')

# Load in the header
header = sc.textFile('../data/CaliforniaHousing/cal_housing.domain')

header.collect()

rdd.take(2)

# Split lines on commas
rdd = rdd.map(lambda line: line.split(","))

# Inspect the first 2 lines 
rdd.take(2)

RDDs are fine for performing low-level transformations and actions on unstructured data. This means that we don't care about imposing a schema or accessing attributed by name or column. We should use RDDs when we want to manipulate data with functional programming constructs rather than domain specific expressions.

### Clean Data and Set Schema

Next, we'll convert this RDD to a DataFrame with the `toDF()` method. This is done by first applying a mapping with each of the desired column names corresponding to a different value in a row.

# Import the necessary modules 
from pyspark.sql import Row

# Map the RDD to a DF
df = rdd.map(lambda line: Row(longitude=line[0], 
                              latitude=line[1], 
                              housingMedianAge=line[2],
                              totalRooms=line[3],
                              totalBedRooms=line[4],
                              population=line[5], 
                              households=line[6],
                              medianIncome=line[7],
                              medianHouseValue=line[8])).toDF()

df.show(5)

df.printSchema()

This data set's schema needs to be amended to assign the correct data types to each column. These are all floating point values and should be converted as such.

# Import all from `sql.types`
from pyspark.sql.types import *

# function to convert the data type of DataFrame columns
def convertColumn(df, names, newType):
  for name in names: 
     df = df.withColumn(name, df[name].cast(newType))
  return df 

# assign all columns to list
columns = df.columns

df = convertColumn(df, columns, FloatType())

df.printSchema()

df.describe().show()

The data is ready to be further analysed and modelled upon. We'll save it in parquet format and continue working on it in another notebook.

df.write.parquet("../data/housing_clean.parquet")

# end spark session for this notebook
sc.stop()
