==> california-housing/notebooks/2-preprocessing-with-pyspark.ipynb <==
# Preprocessing and Feature Engineering
In this notebook, we will further explore the california-housing dataset and prepare the data for modelling. This notebook relies on the output from `1-eda-with-pyspark.ipynb`, which is a parquet file `../data/housing_clean.parquet`.

# Import SparkSession
from pyspark.sql import SparkSession

# Build the SparkSession
spark = SparkSession.builder \
   .master("local") \
   .appName("Linear Regression Model") \
   .config("spark.executor.memory", "1gb") \
   .getOrCreate()
   
sc = spark.sparkContext

df = spark.read.parquet("../data/housing_clean.parquet")

df.show(5)

Check for missing values

from pyspark.sql.functions import *
df.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in df.columns]).show()

Get a summary of the data. This will inform our preprocessing tasks and approach.

df.describe().show()

We can derive the following insights from the data:
- The range of minimum and maximum for the independent variables is quite large, so we should probably standardise the data
- The dependent variable `medianHouseValue` is also quite big, so we should adjust the values
- There are some additional attributes which we could add, such as a feature that counts the number of bedrooms per total room or number of rooms per household

### Preprocessing the Target Values
We'll deal with this in units of 100,000 to shrink the range and make it easier to deal with.

# Adjust the values of `medianHouseValue`
df = df.withColumn("medianHouseValue", col("medianHouseValue")/100000)

df.take(2)

### Feature Engineering
We will add:
- *Rooms per household* number of rooms in households per block group
- *Population per household* how many people live in the households per block group
- *Bedrooms per room* how many rooms are bedrooms per block group

# Divide `totalRooms` by `households`
roomsPerHousehold = df.select(col("totalRooms")/col("households"))

# Divide `population` by `households`
populationPerHousehold = df.select(col("population")/col("households"))

# Divide `totalBedRooms` by `totalRooms`
bedroomsPerRoom = df.select(col("totalBedRooms")/col("totalRooms"))

# Add the new columns to `df`
df = df.withColumn("roomsPerHousehold", col("totalRooms")/col("households")) \
   .withColumn("populationPerHousehold", col("population")/col("households")) \
   .withColumn("bedroomsPerRoom", col("totalBedRooms")/col("totalRooms"))
   
# Inspect the result
df.first()

Now that the features have been created, we want to re-order the columns in the DataFrame so that we isolate the target value before proceeding to standardisation.

# Re-order and select columns
df = df.select("medianHouseValue", 
              "totalBedRooms", 
              "population", 
              "households", 
              "medianIncome", 
              "roomsPerHousehold", 
              "populationPerHousehold", 
              "bedroomsPerRoom")

df.show(5)

### Standardising the Data

There is one more step to go through before normalising the data set; separating the features from the target variable. This boils down to isolating the first column in the DataFrame from the rest of the columns. In this case, we will use the `DenseVector()` function, which creates a local vector that is backed by a double array that represents its entry values. It is what is used to store arrays of values for use in PySpark.

from pyspark.ml.linalg import DenseVector

# Define the `input_data` 
input_data = df.rdd.map(lambda x: (x[0], DenseVector(x[1:])))

# Replace `df` with the new DataFrame
df = spark.createDataFrame(input_data, ["label", "features"])

df.show(5)

Now, we can use the `StandardScaler` from the Spark ML library to easily scale the data.

from pyspark.ml.feature import StandardScaler

# Initialize the `standardScaler`
standardScaler = StandardScaler(inputCol="features", outputCol="features_scaled")

# Fit the DataFrame to the scaler
scaler = standardScaler.fit(df)

# Transform the data in `df` with the scaler
scaled_df = scaler.transform(df)

# Inspect the result
scaled_df.take(2)

Now, the data is clean, feature engineered and normalised, meaning it is ready to model on. We will build a regression model on this data using Spark ML in the next notebook. For now, we will save the data to a parquet file to persist it.

scaled_df.write.parquet("../data/housing_scaled.parquet")

sc.stop()
