==> california-housing/notebooks/3-regression-with-pyspark.ipynb <==
# Machine Learning Model with Spark ML
In this notebook, we build a linear regression model using Spark ML. This notebook relies on the output from `2-preprocessing-with-pyspark.ipynb`, which is a parquet file `../data/housing_scaled.parquet`. 

# Import SparkSession
from pyspark.sql import SparkSession

# Build the SparkSession
spark = SparkSession.builder \
   .master("local") \
   .appName("Linear Regression Model") \
   .config("spark.executor.memory", "1gb") \
   .getOrCreate()
   
sc = spark.sparkContext

scaled_df = spark.read.parquet("../data/housing_scaled.parquet")

scaled_df.show(5)

### Linear Regression Model
We split the data into training and test sets, then train a regularised linear regression model using some naive hyperparameters.

# Split the data into train and test sets
train_data, test_data = scaled_df.randomSplit([.8,.2],seed=1234)

from pyspark.ml.regression import LinearRegression

# Initialize `lr`
lr = LinearRegression(labelCol="label", maxIter=10, regParam=0.3, elasticNetParam=0.8)

# Fit the data to the model
linearModel = lr.fit(train_data)

Next, generate some predictions and compare them to the labels.

# Generate predictions
predicted = linearModel.transform(test_data)

# Extract the predictions and the "known" correct labels
predictions = predicted.select("prediction").rdd.map(lambda x: x[0])
labels = predicted.select("label").rdd.map(lambda x: x[0])

# Zip `predictions` and `labels` into a list
predictionAndLabel = predictions.zip(labels).collect()

# Print out first 5 instances of `predictionAndLabel` 
predictionAndLabel[:5]

sc.stop()
